@@ -7,6 +7,7 @@ from pathlib import Path
 
 from buildingmotif import BuildingMOTIF
 from buildingmotif.dataclasses import Library
+from buildingmotif.ingresses.bacnet import BACnetNetwork
 
 cli = argparse.ArgumentParser(
     prog="buildingmotif", description="CLI Interface for common BuildingMOTIF tasks"
@@ -49,16 +50,6 @@ def get_db_uri(args) -> str:
     db_uri = getenv("DB_URI")
     if db_uri is not None:
         return db_uri
-    try:
-        import configs as building_motif_configs
-    except ImportError:
-        print("No DB URI could be found")
-        print("No configs.py file found")
-        subcommands[args.func].print_help()
-        sys.exit(1)
-    db_uri = building_motif_configs.DB_URI
-    if db_uri is not None:
-        return db_uri
     print("No DB URI could be found")
     subcommands[args.func].print_help()
     sys.exit(1)
@@ -87,7 +78,6 @@ def get_db_uri(args) -> str:
         "--libraries",
         help="Filename of the libraries YAML file specifying what "
         "should be loaded into BuildingMOTIF",
-        default="libraries.yml",
         nargs="+",
         dest="library_manifest_file",
     ),
@@ -97,7 +87,7 @@ def load(args):
     Loads libraries from (1) local directories (--dir),
     (2) local or remote ontology files (--ont)
     (3) library spec file (--libraries): the provided YML file into the
-      BuildingMOTIF instance at $DB_URI or whatever is in 'configs.py'.
+      BuildingMOTIF instance at $DB_URI.
       Use 'get_default_libraries_yml' for the format of the expected libraries.yml file
     """
     db_uri = get_db_uri(args)
@@ -161,6 +151,16 @@ def app():
         cli.print_help()
     else:
         args.func(args)
+
+
+@subcommand(
+    arg("-o", "--output_file", help="Output file for BACnet scan", required=True),
+    arg("-ip", help="ip address of BACnet network to scan", default=None),
+)
+def scan(args):
+    """Scans a BACnet network and generates a JSON file for later processing"""
+    bacnet_network = BACnetNetwork(args.ip)
+    bacnet_network.dump(Path(args.output_file))
 
 
 # entrypoint is actually defined in pyproject.toml; this is here for convenience/testing
