@@ -1,3 +1,5 @@
+from pathlib import Path
+
 from rdflib import Graph, Namespace, URIRef
 from rdflib.compare import isomorphic, to_isomorphic
 from rdflib.namespace import RDF
@@ -25,6 +27,8 @@ default_graph = Graph()
 default_graph.add(
     (URIRef("urn:my_model"), RDF.type, URIRef("http://www.w3.org/2002/07/owl#Ontology"))
 )
+
+PROJECT_DIR = Path(__file__).resolve().parents[3]
 
 
 def test_get_all_models(client, building_motif):
@@ -272,7 +276,11 @@ def test_validate_model(client, building_motif):
 
     assert results.get_json().keys() == {"message", "reasons", "valid"}
     assert isinstance(results.get_json()["message"], str)
-    assert isinstance(results.get_json()["reasons"], list)
+    assert results.get_json()["reasons"] == {
+        "urn:building/vav1": [
+            "urn:building/vav1 needs between 1 and None instances of https://brickschema.org/schema/Brick#Air_Flow_Sensor on path https://brickschema.org/schema/Brick#hasPoint"
+        ]
+    }
     assert not results.get_json()["valid"]
 
     # Set up
@@ -295,7 +303,7 @@ def test_validate_model(client, building_motif):
     assert results.get_json().keys() == {"message", "reasons", "valid"}
     assert isinstance(results.get_json()["message"], str)
     assert results.get_json()["valid"]
-    assert isinstance(results.get_json()["reasons"], list)
+    assert results.get_json()["reasons"] == {}
 
 
 def test_validate_model_bad_model_id(client, building_motif):
@@ -330,7 +338,7 @@ def test_validate_model_no_args(client, building_motif):
     assert results.get_json().keys() == {"message", "reasons", "valid"}
     assert isinstance(results.get_json()["message"], str)
     assert results.get_json()["valid"]
-    assert isinstance(results.get_json()["reasons"], list)
+    assert results.get_json()["reasons"] == {}
 
 
 def test_validate_model_no_library_ids(client, building_motif):
@@ -350,7 +358,7 @@ def test_validate_model_no_library_ids(client, building_motif):
     assert results.get_json().keys() == {"message", "reasons", "valid"}
     assert isinstance(results.get_json()["message"], str)
     assert results.get_json()["valid"]
-    assert isinstance(results.get_json()["reasons"], list)
+    assert results.get_json()["reasons"] == {}
 
 
 def test_validate_model_bad_library_ids(client, building_motif):
@@ -400,3 +408,53 @@ def test_validate_model_bad_args(client, building_motif):
 
     # Assert 2
     assert results.status_code == 400
+
+
+def test_test_model_against_shapes(client, building_motif):
+    # Load libraries
+    Library.load(ontology_graph=str(PROJECT_DIR / "libraries/brick/Brick-subset.ttl"))
+    ashrae_g36 = Library.load(
+        directory=str(PROJECT_DIR / "libraries/ashrae/guideline36/")
+    )
+
+    # build model
+    BLDG = Namespace("http://example.org/building/")
+    medium_office_model = Model.create(BLDG)
+    medium_office_model.graph.parse(
+        PROJECT_DIR
+        / "notebooks/mediumOffice-validation/mediumOffice_brick_compiled.ttl",
+        format="ttl",
+    )
+    manifest = Graph().parse(
+        PROJECT_DIR
+        / "notebooks/mediumOffice-validation/constraints/mediumOffice_constraints.ttl"
+    )
+    medium_office_model.get_manifest().add_graph(manifest)
+
+    # action
+    results = client.post(
+        f"/models/{medium_office_model.id}/validate_shape",
+        headers={"Content-Type": "application/json"},
+        json={
+            "shape_collection_ids": [ashrae_g36.get_shape_collection().id],
+            "shape_uris": [
+                "urn:ashrae/g36/5.16/multiple-zone-vav-air-handling-unit/fc-3",
+                "urn:ashrae/g36/5.16/multiple-zone-vav-air-handling-unit/fc-4",
+            ],
+            "target_class": str(BRICK["AHU"]),
+        },
+    )
+
+    # assert
+    assert (
+        len(
+            results.json["urn:ashrae/g36/5.16/multiple-zone-vav-air-handling-unit/fc-3"]
+        )
+        == 0
+    )
+    assert (
+        len(
+            results.json["urn:ashrae/g36/5.16/multiple-zone-vav-air-handling-unit/fc-4"]
+        )
+        == 3
+    )
